:sparkles: start refactor using iterators
